backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
# psycopg2-binary often has no wheel on very new Python; psycopg v3 works with SQLAlchemy via postgresql+psycopg://
psycopg[binary]==3.3.3
dnspython==2.4.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1
